type fixes
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@4097 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/plugins/imageMAGICK/imageMAGICK.cpp
@@ -52,7 +52,7 @@ imageMAGICK :: imageMAGICK()
   //post("imageMAGICK");
   char**mimelist; 
   char what;
-  unsigned long length;
+  size_t length;
   ExceptionInfo exception;
   GetExceptionInfo(&exception);
   mimelist=GetMimeList("image/*", &length, &exception);
